requirements.txt
aiogram
psycopg2-binary
aiohttp
asyncio